CMS: Split EnvelopedData and SignedData implementations
