+++ src/lib/crypto_wrappers/cms/CMSError.ts
@@ -1,3 +1,3 @@
-import RelaynetError from '../RelaynetError';
+import RelaynetError from '../../RelaynetError';
 
 export default class CMSError extends RelaynetError {}
