+++ src/lib/crypto_wrappers/cms/_test_utils.ts
@@ -1,0 +1,6 @@
+import * as pkijs from 'pkijs';
+import { deserializeDer } from '../_utils';
+
+export function deserializeContentInfo(contentInfoDer: ArrayBuffer): pkijs.ContentInfo {
+  return new pkijs.ContentInfo({ schema: deserializeDer(contentInfoDer) });
+}
