+++ src/lib/ramf/MessageSerializer.spec.ts
@@ -11,7 +11,8 @@ import {
   getMockContext,
   getPromiseRejection,
 } from '../_test_utils';
-import * as cms from '../crypto_wrappers/cms';
+import * as cmsEnvelopedData from '../crypto_wrappers/cms/envelopedData';
+import * as cmsSignedData from '../crypto_wrappers/cms/signedData';
 import { generateRSAKeyPair } from '../crypto_wrappers/keyGenerators';
 import Certificate from '../crypto_wrappers/x509/Certificate';
 import { NON_ASCII_STRING, StubMessage, StubPayload } from './_test_utils';
@@ -342,7 +343,7 @@ describe('MessageSerializer', () => {
     describe('Payload', () => {
       test('Payload should be encrypted', async () => {
         const message = new StubMessage(recipientAddress, senderCertificate, PAYLOAD);
-        jest.spyOn(cms, 'encrypt');
+        jest.spyOn(cmsEnvelopedData, 'encrypt');
 
         const messageSerialized = await STUB_MESSAGE_SERIALIZER.serialize(
           message,
@@ -350,12 +351,16 @@ describe('MessageSerializer', () => {
           recipientCertificate,
         );
 
-        expect(cms.encrypt).toBeCalledTimes(1);
-        expect(cms.encrypt).toBeCalledWith(StubPayload.BUFFER, recipientCertificate, undefined);
+        expect(cmsEnvelopedData.encrypt).toBeCalledTimes(1);
+        expect(cmsEnvelopedData.encrypt).toBeCalledWith(
+          StubPayload.BUFFER,
+          recipientCertificate,
+          undefined,
+        );
 
         const messageParts = parseMessage(messageSerialized);
         const payloadCiphertext = messageParts.payload;
-        const { plaintext } = await cms.decrypt(
+        const { plaintext } = await cmsEnvelopedData.decrypt(
           bufferToArray(payloadCiphertext),
           recipientPrivateKey,
         );
@@ -364,7 +369,7 @@ describe('MessageSerializer', () => {
 
       test('Encryption options should be honoured', async () => {
         const message = new StubMessage(recipientAddress, senderCertificate, PAYLOAD);
-        jest.spyOn(cms, 'encrypt');
+        jest.spyOn(cmsEnvelopedData, 'encrypt');
 
         const encryptionOptions = { aesKeySize: 256 };
         await STUB_MESSAGE_SERIALIZER.serialize(
@@ -374,8 +379,8 @@ describe('MessageSerializer', () => {
           encryptionOptions,
         );
 
-        expect(cms.encrypt).toBeCalledTimes(1);
-        expect(cms.encrypt).toBeCalledWith(
+        expect(cmsEnvelopedData.encrypt).toBeCalledTimes(1);
+        expect(cmsEnvelopedData.encrypt).toBeCalledWith(
           StubPayload.BUFFER,
           recipientCertificate,
           encryptionOptions,
@@ -394,15 +399,15 @@ describe('MessageSerializer', () => {
           senderCertificateChain,
         });
 
-        jest.spyOn(cms, 'sign');
+        jest.spyOn(cmsSignedData, 'sign');
         messageSerialized = await STUB_MESSAGE_SERIALIZER.serialize(
           message,
           senderPrivateKey,
           recipientCertificate,
         );
-        expect(cms.sign).toBeCalledTimes(1);
+        expect(cmsSignedData.sign).toBeCalledTimes(1);
         // @ts-ignore
-        cmsSignArgs = cms.sign.mock.calls[0];
+        cmsSignArgs = cmsSignedData.sign.mock.calls[0];
 
         const messageParts = parseMessage(messageSerialized);
         signature = messageParts.signature;
@@ -447,7 +452,7 @@ describe('MessageSerializer', () => {
         const mockSignature = new ArrayBuffer(0);
         const signatureLength = 2 ** 14;
         jest.spyOn(mockSignature, 'byteLength', 'get').mockReturnValue(signatureLength);
-        jest.spyOn(cms, 'sign').mockReturnValue(Promise.resolve(mockSignature));
+        jest.spyOn(cmsSignedData, 'sign').mockReturnValue(Promise.resolve(mockSignature));
 
         await expectPromiseToReject(
           STUB_MESSAGE_SERIALIZER.serialize(message, senderPrivateKey, recipientCertificate),
@@ -468,13 +473,13 @@ describe('MessageSerializer', () => {
         async hashingAlgorithmName => {
           const message = new StubMessage(recipientAddress, senderCertificate, PAYLOAD);
 
-          jest.spyOn(cms, 'sign');
+          jest.spyOn(cmsSignedData, 'sign');
           await STUB_MESSAGE_SERIALIZER.serialize(message, senderPrivateKey, recipientCertificate, {
             hashingAlgorithmName,
           });
-          expect(cms.sign).toBeCalledTimes(1);
+          expect(cmsSignedData.sign).toBeCalledTimes(1);
           // @ts-ignore
-          const signatureArgs = cms.sign.mock.calls[0];
+          const signatureArgs = cmsSignedData.sign.mock.calls[0];
           expect(signatureArgs[4]).toEqual({ hashingAlgorithmName });
         },
       );
@@ -798,21 +803,21 @@ describe('MessageSerializer', () => {
     describe('Payload', () => {
       test('Payload should be decrypted', async () => {
         const message = new StubMessage(recipientAddress, senderCertificate, PAYLOAD);
-        jest.spyOn(cms, 'encrypt');
+        jest.spyOn(cmsEnvelopedData, 'encrypt');
         const messageSerialized = await STUB_MESSAGE_SERIALIZER.serialize(
           message,
           senderPrivateKey,
           recipientCertificate,
         );
 
-        jest.spyOn(cms, 'decrypt');
+        jest.spyOn(cmsEnvelopedData, 'decrypt');
         const messageDeserialized = await STUB_MESSAGE_SERIALIZER.deserialize(
           messageSerialized,
           recipientPrivateKey,
         );
 
-        const encryptionResult = await getMockContext(cms.encrypt).results[0].value;
-        expect(cms.decrypt).toBeCalledWith(
+        const encryptionResult = await getMockContext(cmsEnvelopedData.encrypt).results[0].value;
+        expect(cmsEnvelopedData.decrypt).toBeCalledWith(
           encryptionResult.envelopedDataSerialized,
           recipientPrivateKey,
         );
@@ -824,23 +829,26 @@ describe('MessageSerializer', () => {
     describe('Signature', () => {
       test('Signature should be accepted if valid', async () => {
         const message = new StubMessage(recipientAddress, senderCertificate, PAYLOAD);
-        jest.spyOn(cms, 'sign');
+        jest.spyOn(cmsSignedData, 'sign');
         const messageSerialized = await STUB_MESSAGE_SERIALIZER.serialize(
           message,
           senderPrivateKey,
           recipientCertificate,
         );
 
-        jest.spyOn(cms, 'verifySignature');
+        jest.spyOn(cmsSignedData, 'verifySignature');
         await STUB_MESSAGE_SERIALIZER.deserialize(messageSerialized, recipientPrivateKey);
 
-        const signatureCiphertext = await getMockContext(cms.sign).results[0].value;
+        const signatureCiphertext = await getMockContext(cmsSignedData.sign).results[0].value;
         const signaturePlaintext = messageSerialized.slice(
           0,
           messageSerialized.byteLength - signatureCiphertext.length - 2,
         );
-        expect(cms.verifySignature).toBeCalledTimes(1);
-        expect(cms.verifySignature).toBeCalledWith(signatureCiphertext, signaturePlaintext);
+        expect(cmsSignedData.verifySignature).toBeCalledTimes(1);
+        expect(cmsSignedData.verifySignature).toBeCalledWith(
+          signatureCiphertext,
+          signaturePlaintext,
+        );
       });
 
       test('Signature should not be accepted if invalid', async () => {
@@ -848,7 +856,7 @@ describe('MessageSerializer', () => {
         const signerCertificate = await generateStubCert({
           subjectPublicKey: signerKeyPair.publicKey,
         });
-        const invalidSignature = await cms.sign(
+        const invalidSignature = await cmsSignedData.sign(
           bufferToArray(Buffer.from('Hello world')),
           signerKeyPair.privateKey,
           signerCertificate,
@@ -871,7 +879,7 @@ describe('MessageSerializer', () => {
         const signerCertificate = await generateStubCert({
           subjectPublicKey: signerKeyPair.publicKey,
         });
-        const invalidSignature = await cms.sign(
+        const invalidSignature = await cmsSignedData.sign(
           bufferToArray(Buffer.from('Hello world')),
           signerKeyPair.privateKey,
           signerCertificate,
@@ -907,12 +915,12 @@ describe('MessageSerializer', () => {
 
       test('Sender certificate chain should be extracted from signature', async () => {
         const caCertificate = await generateStubCert();
-        const encryptionResult = await cms.encrypt(PAYLOAD, recipientCertificate);
+        const encryptionResult = await cmsEnvelopedData.encrypt(PAYLOAD, recipientCertificate);
         const messageSerialized = await serializeWithoutValidation({
           payloadBuffer: encryptionResult.envelopedDataSerialized,
         });
 
-        jest.spyOn(cms, 'verifySignature').mockImplementationOnce(async () => ({
+        jest.spyOn(cmsSignedData, 'verifySignature').mockImplementationOnce(async () => ({
           signerCertificate: senderCertificate,
           signerCertificateChain: [senderCertificate, caCertificate],
         }));
@@ -974,7 +982,7 @@ describe('MessageSerializer', () => {
 
     const finalSignature = Buffer.from(
       signature ||
-        (await cms.sign(
+        (await cmsSignedData.sign(
           bufferToArray(serialization.toBuffer()),
           senderPrivateKey,
           senderCertificate,
