+++ src/lib/ramf/MessageSerializer.ts
@@ -2,7 +2,8 @@ import { Parser } from 'binary-parser';
 import bufferToArray from 'buffer-to-arraybuffer';
 import { SmartBuffer } from 'smart-buffer';
 
-import * as cms from '../crypto_wrappers/cms';
+import * as cmsEnvelopedData from '../crypto_wrappers/cms/envelopedData';
+import * as cmsSignedData from '../crypto_wrappers/cms/signedData';
 import Certificate from '../crypto_wrappers/x509/Certificate';
 import Message from './Message';
 import RAMFSyntaxError from './RAMFSyntaxError';
@@ -64,7 +65,7 @@ export class MessageSerializer<MessageSpecialization extends Message> {
     message: MessageSpecialization,
     senderPrivateKey: CryptoKey,
     recipientCertificate: Certificate,
-    options?: Partial<cms.EncryptionOptions | cms.SignatureOptions>,
+    options?: Partial<cmsEnvelopedData.EncryptionOptions | cmsSignedData.SignatureOptions>,
   ): Promise<ArrayBuffer> {
     //region Validation
     validateRecipientAddressLength(message.recipientAddress);
@@ -103,10 +104,10 @@ export class MessageSerializer<MessageSpecialization extends Message> {
     //endregion
 
     //region Payload
-    const { envelopedDataSerialized } = await cms.encrypt(
+    const { envelopedDataSerialized } = await cmsEnvelopedData.encrypt(
       message.exportPayload(),
       recipientCertificate,
-      options as cms.EncryptionOptions,
+      options as cmsEnvelopedData.EncryptionOptions,
     );
     serialization.writeUInt32LE(envelopedDataSerialized.byteLength);
     serialization.writeBuffer(Buffer.from(envelopedDataSerialized));
@@ -115,12 +116,12 @@ export class MessageSerializer<MessageSpecialization extends Message> {
     const serializationBeforeSignature = serialization.toBuffer();
 
     //region Signature
-    const signature = await cms.sign(
+    const signature = await cmsSignedData.sign(
       bufferToArray(serializationBeforeSignature),
       senderPrivateKey,
       message.senderCertificate,
       new Set([message.senderCertificate, ...message.senderCertificateChain]),
-      options as cms.SignatureOptions,
+      options as cmsSignedData.SignatureOptions,
     );
     validateSignatureLength(signature);
     const signatureLengthPrefix = Buffer.allocUnsafe(2);
@@ -153,7 +154,7 @@ export class MessageSerializer<MessageSpecialization extends Message> {
     validateMessageTiming(messageFields, signatureVerification);
     //endregion
 
-    const decryptionResult = await cms.decrypt(
+    const decryptionResult = await cmsEnvelopedData.decrypt(
       bufferToArray(messageFields.payload),
       recipientPrivateKey,
     );
@@ -259,7 +260,7 @@ function parseMessage(serialization: ArrayBuffer): MessageFields {
 async function verifySignature(
   messageSerialized: ArrayBuffer,
   messageFields: MessageFields,
-): Promise<cms.SignatureVerification> {
+): Promise<cmsSignedData.SignatureVerification> {
   const signatureCiphertext = bufferToArray(messageFields.signature);
   const signatureCiphertextLengthWithLengthPrefix = 2 + signatureCiphertext.byteLength;
   const signaturePlaintext = messageSerialized.slice(
@@ -267,7 +268,7 @@ async function verifySignature(
     messageSerialized.byteLength - signatureCiphertextLengthWithLengthPrefix,
   );
   try {
-    return await cms.verifySignature(signatureCiphertext, signaturePlaintext);
+    return await cmsSignedData.verifySignature(signatureCiphertext, signaturePlaintext);
   } catch (error) {
     throw new RAMFValidationError('Invalid RAMF message signature', messageFields, error);
   }
@@ -275,7 +276,7 @@ async function verifySignature(
 
 function validateMessageTiming(
   messageFields: MessageFields,
-  signatureVerification: cms.SignatureVerification,
+  signatureVerification: cmsSignedData.SignatureVerification,
 ): void {
   const currentTimestamp = dateToTimestamp(new Date());
   if (currentTimestamp < messageFields.dateTimestamp) {
